Create the sign up template and write its logic
The signup form template was created. Validation rules were added so
if the data entered by the user pass the validation rules, the user will
be saved to the database and they will be redirected to the homepage. If
not, the form would rerender and the form fields would be populated with
the previous data and the error messages at the bottom of the page.

Passwords would be saved to the database as hashed values.