Prepare the project for production and fix bugs
Install packages such as helmet, compression, express-rate-limit to
manage the aspects of the app that were not crucial in development mode.
Fixed a escaping bug. Removed escaping sanitization from the user and
post validators to store the raw data in the database and instead manage
the escaping on the client-side wherever necessary.